# bootstrap/format: send larger batches to rustfmt