@Urgau 2024-02-29
Did you try with [`std::thread::available_parallelism`](https://doc.rust-lang.org/std/thread/fn.available_parallelism.html) ? Did it make a difference ?