@RalfJung 2024-02-29
This code quite deliberately honors `-j`, I think.

On my system, available_parallelism is 20, so that just means we're no longer measuring what happens on low-core systems.