@the8472 2024-02-29
> On my system, available_parallelism is 20, so that just means we're no longer measuring what happens on low-core systems.

The batch size was set to a small value because it might end up distributing work unevenly across rustfmt processes and thus lead to worse wall-time on many-core systems.
Have you measured the impact without `-j 1`?