@the8472 2024-02-29
Actually, how did you benchmark this? I'm currently looking at the bootstrap rustfmt code and its behavior depends on the git state.
If you're on a clean master then it checks everything.
If you're on a branch or you have modified any files it only checks modified files.

This can taint benchmark results.