@RalfJung 2024-02-29
Please read the PR description, I described my benchmarking methodology. :)